rename: Bookmark -> BookmarkPage로 수정
